support for split sections
not very useful yet since there's no referencing or table-of-contents,
but it's a start.

sections no longer return the section, and instead just append a child.
inline sections will always be appended in order to the parent's
content. this means there can't be any content after the inlined
sections, but that's very unintuitive to read anyway.

NB: the Write stage is a bit silly now. the first 'if' never triggers.
it should really just not be a visitor pattern.