change how sub-sections work so they can be tagged
plugins can now defer evaluation of their arguments, with the caveat
that they need to be able to evaluate it, i.e. by having the processor.

i don't know if this will be a super common thing for plugins, so for
now there's just a delicate initialization dance between the processor
and the baselit plugin, since they're inter-dependent.

as a result, though, sections now just use \tile{Foo}{foo} like normal,
which is a bit less succinct but it does make it easier to copy sections
into separate files.