add \table-of-contents
* render nested inline sections directly in the template via
recursion rather than in the writer, which would only nest one level.
* add section numbers to headings after the top level. this will likely
  be configurable in the future.
* fix syntax of indented invoke arguments (just needed to skip
  whitespace leading up to close brace)

one concern I have with the current approach is that TableOfContents is
purely a rendering-time concern and has no bearing on the "content", per
se, and yet was added as a Content type. if more of these come up it may
be worth revisiting as a general 'RenderContent' thing, so that Content
visitors like 'Resolve' (and later 'Collect') can more easily skip them.