add --plugin flag
ok, hear me out. this is pretty wackadoo.

when --plugins are specified, booklit generates a main.go on-the-fly
with each plugin under-score imported, and then builds and re-executes
*that* binary.

the idea here is if you're using plugins you may be developing them
alongside your docs anyway, so it's not such a crazy assumption that
you'd have a Go toolchain available.