add support for section partials
this allows sections to retain named partials that can be rendered at
specific points in the template by name. it effectively introduces a
data transfer route to and from the document, its plugins, and the
template.

this can fix the ugly pattern of having plugins with giant methods that
have to exist to generate "html" (via Block/Element and tons of classes)
because they intersperse content arguments into it.  plugins can instead
set specific named partials for the content, and utilize the rendering
engine (plus the html engine's templating) to do all the html-specific
things by accessing the partials.