resolve ambiguity with table/definitions items
previously if only a single row/entry was given, it would be the
detected sequence.

now, an arg is given for each entry, which may aid in readability in the
long run anyway, as it'll force entries to be spaced apart.

\definition and \table-row are aliases to \list so that there's an
explicit structure to detect rather than just Sequence/Paragraph