check against ambiguous references
prior to this change it was way too easy to define the same tag in
multiple sections (either up/down the hierarchy or in sibling sections)
and have a \reference point to the wrong tag.

this will now behave similar to a broken reference: an error will be
raised enumerating all of the sections that define the tag, so you can
go make them more explicitly named.

passing --allow-broken-references will suppress the error. also, all
broken references (ambiguous or missing) are now logged, so you can keep
track of them while you whittle away.

fixes #14