screw this
I think Booklit syntax is just a bit too flexible for extensible
Markdown parsers. This isn't really a knock on Goldmark, but it doesn't
seem to be possible (or at least easy) to support this form:

    \foo{bar}{
      baz
    }

It supports block parsing and inline parsing, but the above example
demonstrates both: I want to parse `bar` inline, and `baz` as contents
of the overall block.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>