checkpoint: tests passing with janky block args
a compromise: since it doesn't seem possible to handle this syntax...:

  \foo{bar}{
    baz
  }

...because of the Markdown block vs. inline parsing approach, where
block start/end lines can't coexist with inline content, let's try
to handle this syntax instead:

  \foo{bar}
  {
    baz
  }

This can be done by having a separate block parser which returns a
'block argument' node which we then stitch back into the Invoke node
while traversing the tree.

It's a little janky, but the parsing logic is a whole lot simpler -
there's just an inline arg parser and a block arg parser.

Might even be able to achieve the original syntax with some really
simple pre-processing that just splits the '{' onto the next line.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>