don't match section title for \reference
not sure why this was implemented - it's not documented, but there's a
test for it.

i think it's better for \reference to always point to a tag. using the
title directly makes the documentation more brittle, because if I change
the section's title I break all the references *and* any external
references to the page.

using tags instead encourages users to come up with short, stable tag
names for sections to prevent URL breakage.

also this would result in an AmbiguousReferenceError if a title was the
same as the computed tag, e.g. a single lowercase word. that could have
been fixed, but this seems like a better approach.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>